hcm: factoring out filter chain creation code
