+++ source/common/http/dependency_manager.cc
@@ -1,12 +1,10 @@
-#include "source/extensions/filters/network/http_connection_manager/dependency_manager.h"
+#include "source/common/http/dependency_manager.h"
 
 #include "absl/container/flat_hash_set.h"
 #include "absl/strings/substitute.h"
 
 namespace Envoy {
-namespace Extensions {
-namespace NetworkFilters {
-namespace HttpConnectionManager {
+namespace Http {
 
 using envoy::extensions::filters::common::dependency::v3::Dependency;
 
@@ -30,7 +28,5 @@ absl::Status DependencyManager::validDecodeDependencies() {
   return absl::OkStatus();
 }
 
-} // namespace HttpConnectionManager
-} // namespace NetworkFilters
-} // namespace Extensions
+} // namespace Http
 } // namespace Envoy
