+++ source/extensions/filters/network/http_connection_manager/BUILD
@@ -1,7 +1,6 @@
 load(
     "//bazel:envoy_build_system.bzl",
     "envoy_cc_extension",
-    "envoy_cc_library",
     "envoy_extension_package",
     "envoy_select_enable_http3",
 )
@@ -21,7 +20,6 @@ envoy_cc_extension(
     # This is core Envoy config.
     visibility = ["//visibility:public"],
     deps = [
-        ":dependency_manager",
         "//envoy/config:config_provider_manager_interface",
         "//envoy/filesystem:filesystem_interface",
         "//envoy/http:codec_interface",
@@ -41,6 +39,7 @@ envoy_cc_extension(
         "//source/common/http:default_server_string_lib",
         "//source/common/http:request_id_extension_lib",
         "//source/common/http:utility_lib",
+        "//source/common/http:filter_chain_helper_lib",
         "//source/common/http/http1:codec_lib",
         "//source/common/http/http1:settings_lib",
         "//source/common/http/http2:codec_lib",
@@ -56,7 +55,6 @@ envoy_cc_extension(
         "//source/common/tracing:tracer_config_lib",
         "//source/common/tracing:custom_tag_lib",
         "//source/common/network:cidr_range_lib",
-        "//source/extensions/filters/http/common:pass_through_filter_lib",
         "//source/extensions/filters/network:well_known_names",
         "//source/extensions/filters/network/common:factory_base_lib",
         "//source/extensions/http/original_ip_detection/xff:config",
@@ -69,16 +67,4 @@ envoy_cc_extension(
     ] + envoy_select_enable_http3([
         "//source/common/quic:codec_lib",
     ]),
-)
-
-envoy_cc_library(
-    name = "dependency_manager",
-    srcs = ["dependency_manager.cc"],
-    hdrs = ["dependency_manager.h"],
-    external_deps = [
-        "abseil_status",
-    ],
-    deps = [
-        "@envoy_api//envoy/extensions/filters/common/dependency/v3:pkg_cc_proto",
-    ],
 )
