+++ source/extensions/filters/network/http_connection_manager/config.cc
@@ -23,6 +23,7 @@
 #include "source/common/http/conn_manager_config.h"
 #include "source/common/http/conn_manager_utility.h"
 #include "source/common/http/default_server_string.h"
+#include "source/common/http/filter_chain_helper.h"
 #include "source/common/http/http1/codec_impl.h"
 #include "source/common/http/http1/settings.h"
 #include "source/common/http/http2/codec_impl.h"
@@ -40,8 +41,6 @@
 #ifdef ENVOY_ENABLE_QUIC
 #include "source/common/quic/codec_impl.h"
 #endif
-
-#include "source/extensions/filters/http/common/pass_through_filter.h"
 
 namespace Envoy {
 namespace Extensions {
@@ -82,21 +81,6 @@ std::unique_ptr<Http::InternalAddressConfig> createInternalAddressConfig(
 
   return std::make_unique<Http::DefaultInternalAddressConfig>();
 }
-
-class MissingConfigFilter : public Http::PassThroughDecoderFilter {
-public:
-  Http::FilterHeadersStatus decodeHeaders(Http::RequestHeaderMap&, bool) override {
-    decoder_callbacks_->streamInfo().setResponseFlag(StreamInfo::ResponseFlag::NoFilterConfigFound);
-    decoder_callbacks_->sendLocalReply(Http::Code::InternalServerError, EMPTY_STRING, nullptr,
-                                       absl::nullopt, EMPTY_STRING);
-    return Http::FilterHeadersStatus::StopIteration;
-  }
-};
-
-static Http::FilterFactoryCb MissingConfigFilterFactory =
-    [](Http::FilterChainFactoryCallbacks& cb) {
-      cb.addStreamDecoderFilter(std::make_shared<MissingConfigFilter>());
-    };
 
 envoy::extensions::filters::network::http_connection_manager::v3::HttpConnectionManager::
     PathWithEscapedSlashesAction
@@ -150,7 +134,6 @@ SINGLETON_MANAGER_REGISTRATION(date_provider);
 SINGLETON_MANAGER_REGISTRATION(route_config_provider_manager);
 SINGLETON_MANAGER_REGISTRATION(scoped_routes_config_provider_manager);
 SINGLETON_MANAGER_REGISTRATION(http_tracer_manager);
-SINGLETON_MANAGER_REGISTRATION(filter_config_provider_manager);
 
 Utility::Singletons Utility::createSingletons(Server::Configuration::FactoryContext& context) {
   std::shared_ptr<Http::TlsCachingDateProviderImpl> date_provider =
@@ -182,9 +165,8 @@ Utility::Singletons Utility::createSingletons(Server::Configuration::FactoryCont
       });
 
   std::shared_ptr<FilterConfigProviderManager> filter_config_provider_manager =
-      context.singletonManager().getTyped<FilterConfigProviderManager>(
-          SINGLETON_MANAGER_REGISTERED_NAME(filter_config_provider_manager),
-          [] { return std::make_shared<Filter::HttpFilterConfigProviderManagerImpl>(); });
+      Http::FilterChainHelper::createSingletonFilterConfigProviderManager(
+          context.getServerFactoryContext());
 
   return {date_provider, route_config_provider_manager, scoped_routes_config_provider_manager,
           http_tracer_manager, filter_config_provider_manager};
@@ -579,17 +561,8 @@ HttpConnectionManagerConfig::HttpConnectionManagerConfig(
     throw EnvoyException("Non-HTTP/3 codec configured on QUIC listener.");
   }
 
-  const auto& filters = config.http_filters();
-  DependencyManager dependency_manager;
-  for (int32_t i = 0; i < filters.size(); i++) {
-    processFilter(filters[i], i, "http", "http", i == filters.size() - 1, filter_factories_,
-                  dependency_manager);
-  }
-  // TODO(auni53): Validate encode dependencies too.
-  auto status = dependency_manager.validDecodeDependencies();
-  if (!status.ok()) {
-    throw EnvoyException(std::string(status.message()));
-  }
+  Http::FilterChainHelper helper(filter_config_provider_manager_, context_, stats_prefix_);
+  helper.processFilters(config.http_filters(), "http", "http", filter_factories_);
 
   for (const auto& upgrade_config : config.upgrade_configs()) {
     const std::string& name = upgrade_config.upgrade_type();
@@ -601,12 +574,8 @@ HttpConnectionManagerConfig::HttpConnectionManagerConfig(
     }
     if (!upgrade_config.filters().empty()) {
       std::unique_ptr<FilterFactoriesList> factories = std::make_unique<FilterFactoriesList>();
-      DependencyManager upgrade_dependency_manager;
-      for (int32_t j = 0; j < upgrade_config.filters().size(); j++) {
-        processFilter(upgrade_config.filters(j), j, name, "http upgrade",
-                      j == upgrade_config.filters().size() - 1, *factories,
-                      upgrade_dependency_manager);
-      }
+      Http::DependencyManager upgrade_dependency_manager;
+      helper.processFilters(upgrade_config.filters(), name, "http upgrade", *factories);
       // TODO(auni53): Validate encode dependencies too.
       auto status = upgrade_dependency_manager.validDecodeDependencies();
       if (!status.ok()) {
@@ -621,74 +590,6 @@ HttpConnectionManagerConfig::HttpConnectionManagerConfig(
           std::make_pair(name, FilterConfig{std::move(factories), enabled}));
     }
   }
-}
-
-void HttpConnectionManagerConfig::processFilter(
-    const envoy::extensions::filters::network::http_connection_manager::v3::HttpFilter&
-        proto_config,
-    int i, const std::string& prefix, const std::string& filter_chain_type,
-    bool last_filter_in_current_config, FilterFactoriesList& filter_factories,
-    DependencyManager& dependency_manager) {
-  ENVOY_LOG(debug, "    {} filter #{}", prefix, i);
-  if (proto_config.config_type_case() ==
-      envoy::extensions::filters::network::http_connection_manager::v3::HttpFilter::ConfigTypeCase::
-          kConfigDiscovery) {
-    processDynamicFilterConfig(proto_config.name(), proto_config.config_discovery(),
-                               filter_factories, filter_chain_type, last_filter_in_current_config);
-    return;
-  }
-
-  // Now see if there is a factory that will accept the config.
-  auto* factory =
-      Config::Utility::getAndCheckFactory<Server::Configuration::NamedHttpFilterConfigFactory>(
-          proto_config, proto_config.is_optional());
-  // null pointer returned only when the filter is optional, then skip all the processes.
-  if (factory == nullptr) {
-    ENVOY_LOG(warn, "Didn't find a registered factory for the optional http filter {}",
-              proto_config.name());
-    return;
-  }
-  ProtobufTypes::MessagePtr message = Config::Utility::translateToFactoryConfig(
-      proto_config, context_.messageValidationVisitor(), *factory);
-  Http::FilterFactoryCb callback =
-      factory->createFilterFactoryFromProto(*message, stats_prefix_, context_);
-  dependency_manager.registerFilter(factory->name(), *factory->dependencies());
-  bool is_terminal = factory->isTerminalFilterByProto(*message, context_);
-  Config::Utility::validateTerminalFilters(proto_config.name(), factory->name(), filter_chain_type,
-                                           is_terminal, last_filter_in_current_config);
-  auto filter_config_provider = filter_config_provider_manager_.createStaticFilterConfigProvider(
-      {factory->name(), callback}, proto_config.name());
-  ENVOY_LOG(debug, "      name: {}", filter_config_provider->name());
-  ENVOY_LOG(debug, "    config: {}",
-            MessageUtil::getJsonStringFromMessageOrError(
-                static_cast<const Protobuf::Message&>(proto_config.typed_config())));
-  filter_factories.push_back(std::move(filter_config_provider));
-}
-
-void HttpConnectionManagerConfig::processDynamicFilterConfig(
-    const std::string& name, const envoy::config::core::v3::ExtensionConfigSource& config_discovery,
-    FilterFactoriesList& filter_factories, const std::string& filter_chain_type,
-    bool last_filter_in_current_config) {
-  ENVOY_LOG(debug, "      dynamic filter name: {}", name);
-  if (config_discovery.apply_default_config_without_warming() &&
-      !config_discovery.has_default_config()) {
-    throw EnvoyException(fmt::format(
-        "Error: filter config {} applied without warming but has no default config.", name));
-  }
-  for (const auto& type_url : config_discovery.type_urls()) {
-    auto factory_type_url = TypeUtil::typeUrlToDescriptorFullName(type_url);
-    auto* factory = Registry::FactoryRegistry<
-        Server::Configuration::NamedHttpFilterConfigFactory>::getFactoryByType(factory_type_url);
-    if (factory == nullptr) {
-      throw EnvoyException(
-          fmt::format("Error: no factory found for a required type URL {}.", factory_type_url));
-    }
-  }
-
-  auto filter_config_provider = filter_config_provider_manager_.createDynamicFilterConfigProvider(
-      config_discovery, name, context_, stats_prefix_, last_filter_in_current_config,
-      filter_chain_type, nullptr);
-  filter_factories.push_back(std::move(filter_config_provider));
 }
 
 Http::ServerConnectionPtr
@@ -728,31 +629,8 @@ HttpConnectionManagerConfig::createCodec(Network::Connection& connection,
   PANIC_DUE_TO_CORRUPT_ENUM;
 }
 
-void HttpConnectionManagerConfig::createFilterChainForFactories(
-    Http::FilterChainManager& manager, const FilterFactoriesList& filter_factories) {
-  bool added_missing_config_filter = false;
-  for (const auto& filter_config_provider : filter_factories) {
-    auto config = filter_config_provider->config();
-    if (config.has_value()) {
-      Filter::NamedHttpFilterFactoryCb& factory_cb = config.value().get();
-      manager.applyFilterFactoryCb({filter_config_provider->name(), factory_cb.name},
-                                   factory_cb.factory_cb);
-      continue;
-    }
-
-    // If a filter config is missing after warming, inject a local reply with status 500.
-    if (!added_missing_config_filter) {
-      ENVOY_LOG(trace, "Missing filter config for a provider {}", filter_config_provider->name());
-      manager.applyFilterFactoryCb({}, MissingConfigFilterFactory);
-      added_missing_config_filter = true;
-    } else {
-      ENVOY_LOG(trace, "Provider {} missing a filter config", filter_config_provider->name());
-    }
-  }
-}
-
 void HttpConnectionManagerConfig::createFilterChain(Http::FilterChainManager& manager) {
-  createFilterChainForFactories(manager, filter_factories_);
+  Http::FilterChainHelper::createFilterChainForFactories(manager, filter_factories_);
 }
 
 bool HttpConnectionManagerConfig::createUpgradeFilterChain(
@@ -783,7 +661,7 @@ bool HttpConnectionManagerConfig::createUpgradeFilterChain(
     filters_to_use = it->second.filter_factories.get();
   }
 
-  createFilterChainForFactories(callbacks, *filters_to_use);
+  Http::FilterChainHelper::createFilterChainForFactories(callbacks, *filters_to_use);
   return true;
 }
