+++ test/extensions/filters/network/http_connection_manager/BUILD
@@ -1,6 +1,5 @@
 load(
     "//bazel:envoy_build_system.bzl",
-    "envoy_cc_test",
     "envoy_package",
     "envoy_proto_library",
 )
@@ -83,13 +82,5 @@ envoy_extension_cc_test(
         ":config_test_base",
         "//source/extensions/filters/network/http_connection_manager:config",
         "@envoy_api//envoy/extensions/filters/common/dependency/v3:pkg_cc_proto",
-    ],
-)
-
-envoy_cc_test(
-    name = "dependency_manager_test",
-    srcs = ["dependency_manager_test.cc"],
-    deps = [
-        "//source/extensions/filters/network/http_connection_manager:dependency_manager",
     ],
 )
